Moves logic to QueueMessageProcessor.cs + makes sure that processing is async
